==> 1.2 Prepare mapping file (Python).ipynb <==
This notebook is to add additional covariates (health, stool info, medication, dietary history) to the mapping file.

import pandas as pd
import numpy as np

mf = pd.read_csv('../data/mapping_MrOS.txt', sep='\t', dtype=str, index_col='#SampleID') 

print(mf.shape)
mf.head()

## convert covariates to proper continuous or categorical format

vars_cat = np.array(['BarcodeSequence', 'LinkerPrimerSequence', 'Experiment_Design_Description',
             'Library_Construction_Protocol', 'Linker', 'Platform', 'Center_Name', 'Center_Project', 'Instrument_Model',
             'Title', 'Anonymized_Name', 'Scientific_Name', 'Taxon_ID', 'Sample_Type', 'Geo_Loc_Name', 'Elevation', 'Env_Biome',
             'Env_Feature', 'Env_Material', 'Env_Package', 'Collection_Timestamp', 'DNA_Extracted', 'Physical_Specimen_Location',
             'Physical_Specimen_Remaining', 'Age_Units', 'Host_Subject_ID', 'Host_Taxid','Host_Scientific_Name', 'Host_Common_Name',
             'Life_Stage', 'Sex', 'Height_Units', 'Weight_Units', 'Body_Habitat', 'Body_Site', 'Body_Product', 'GIERACE', 'SITE',
             'TUDRAMT', 'TURSMOKE', 'M1ADEPR', 'M1VITMND', 'M1ANTIB', 'M1PROBI', 'OHSEAS', 'VDstatus', 'Description',
             'OHV1D2CT', 'OHVD2CT'])
vars_cts = np.array(['Latitude', 'Longitude', 'Age', 'Height', 'Weight', 'BMI', 'PASCORE', 'DTVITD', 
             'OHV1D3', 'OHV24D3', 'OHVD3', 'OHVD2', 'OHV1D2',  'OHVDTOT', 'OHV1DTOT'])

# convert vars_cts to numeric and vars_cat to factors
df = mf.copy()
df[vars_cts] = df[vars_cts].apply(pd.to_numeric, errors='coerce')
df[vars_cat] = df[vars_cat].apply(lambda x: x.astype('category'))

# convert all pg/ml to ng/ml note: 1 ng/ml = 1000 pg/ml
df.OHV1D3 = df.OHV1D3/1000
df.OHV1D2 = df.OHV1D2/1000
df.OHV1DTOT = df.OHV1DTOT/1000

## add two ratio variables of Vitamin D

df['ratio_activation'] = df.OHV1D3/df.OHVD3
df['ratio_catabolism'] = df.OHV24D3/df.OHVD3 
vars_cts = np.append(vars_cts, ['ratio_activation', 'ratio_catabolism'])

df[vars_cts].describe()

df[vars_cat].describe()

# count missing values
df[vars_cts].isnull().sum()

# check
df.shape

## add health variables

health = pd.read_csv('../data/MrOS_healthvari.csv', sep=',', dtype=str)
health = health.rename(columns={'ID': '#SampleID'}).set_index('#SampleID')
health['QLCOMP'] = health['QLCOMP'].astype('category')

df = pd.merge(df, health, left_index=True, right_index=True)

print(df.shape)
df.head()

## add stool sample variables

stool = pd.read_csv('../data/Stool600comments_JS.txt', sep='\t', dtype=str)
stool = stool.rename(columns={'ID': '#SampleID'}).set_index('#SampleID')

id_stool = list(stool.index)
id_mapping = list(df.index)

for i in id_mapping:
    if i not in id_stool:
        print(i)
        
for i in id_stool:
    if i not in id_mapping:
        print(i)

stool['Timestool '] = stool['Timestool '].astype('category') # notice a space in name
stool['Timestoolc'] = stool['Timestoolc'].astype('category')
stool['StoolCollected'] = stool['StoolCollected'].astype('category')
stool['StoolShipped'] = stool['StoolShipped'].astype('category')
stool['Days( between collection and shipment)'] = stool['Days( between collection and shipment)'].astype('category')
stool['StoolQualityC'] = stool['StoolQualityC'].astype('category')

## add medication variables

med = pd.read_csv('../data/Microbiom_medsupdate.csv', sep=',', dtype=str)

med.head()

med = med.rename(columns={'ID': '#SampleID'}).set_index('#SampleID')
med = med.apply(lambda x: x.astype('category'))

med.head()

df = pd.merge(df, med, left_index=True, right_index=True)

df.shape

df.head()

## add starch variables

st = pd.read_csv('../data/Microbiome_DietaryStarchesnew.csv', sep=',')

st.shape

st = st.rename(columns={'ID': '#SampleID', 'Rstarches_c (0: <5 and 1>=5)': 'Rstarches_c'}).set_index('#SampleID')

st.head()

df = pd.merge(df, st, left_index=True, right_index=True)

df.shape

### add PPI medication information

ppi = pd.read_csv('../data/PPI_info_mros.txt', sep='\t')

ppi.shape

ppi = ppi.rename(columns={'ID': '#SampleID'}).set_index('#SampleID')

ppi.head()

ppi = ppi.drop(columns=['MIDATA'])
df = pd.merge(df, ppi, left_index=True, right_index=True)

df.shape

## output new mapping file

df.to_csv('../data/mapping_MrOS_add.txt', sep='\t', index_label='#SampleID')

df.head()
