==> 2.1 Alpha Diversity Analysis -- Add to mapping file (Python).ipynb <==
#### This notebook is to add alpha diversities to existing mapping file

import pandas as pd
import numpy as np

mf = pd.read_csv('../data/mapping_MrOS_add.txt', sep='\t', dtype=str, index_col='#SampleID')

mf.head()

mf.shape

## add rarefied PD alpha diversity

alpha_PD = pd.read_csv('../data/57316-alpha_PD/alpha-diversity.tsv', sep='\t', index_col=0)

alpha_PD.head()

alpha_PD.rename(columns={'faith_pd': 'alpha_pd'}, inplace=True)

df = pd.merge(mf, alpha_PD, left_index=True, right_index=True)

df.shape

df.head()

## export

df.to_csv('../data/mros_mapping_alpha.txt', sep='\t', index_label='#SampleID')
