==> 3.1 Beta Diversity (UW) -- RDA (R).ipynb <==
#### This notebook is to perform Non-Redundant Analysis on beta diversities

##### reference: Falony G, Joossens M, Vieira-Silva S, et al. Population-level analysis of gut microbiome variation. Science 2016;352:560-4

library(vegan)

mf = read.csv('../data/mros_mapping_alpha.txt', stringsAsFactors=FALSE, sep='\t')
colnames(mf)[1] = 'SampleName'
dim(mf)

pc = read.csv('../data/57316-unweighted_unifrac_pcoa_results/ordination.txt', 
              header=FALSE, sep='\t', skip=9)

head(pc)

pc_10 = pc[, 1:11]
colnames(pc_10) = c('SampleName', 'PC1', 'PC2', 'PC3', 'PC4', 'PC5', 'PC6', 'PC7', 'PC8', 'PC9', 'PC10')
dim(pc_10)
head(pc_10)

head(mf)

dat = merge(mf, pc_10, by='SampleName')
dim(dat)

head(dat)

# not include correlated variables
# convert categorial to factors
vars_cat = c('Sex','GIERACE', 'SITE', 'TUDRAMT', 'TURSMOKE', 'M1ADEPR', 'M1VITMND', 'M1ANTIB', 'M1PROBI', 
               'OHSEAS', 'QLCOMP', 'VDstatus', 'M1STATIN', 'Antihistamine', 'Laxative', 'M1PPUMP')
dat[vars_cat] = lapply(dat[vars_cat], factor)

# convert continuous to numeric
vars_cts = c('Latitude', 'Longitude', 'Age', 'BMI', 'PASCORE', 'DTVITD', 'Rstarch_total',
             'OHV1D3', 'OHV24D3', 'OHVD3')
vars_pcs = c('PC1', 'PC2', 'PC3', 'PC4', 'PC5', 'PC6', 'PC7', 'PC8', 'PC9', 'PC10')

length(vars_cat)
length(vars_cts)
length(vars_pcs)

dat = dat[c(vars_cat, vars_cts, vars_pcs)]
dat = dat[complete.cases(dat), ]
X = dat[c(vars_cat, vars_cts)]
Y = dat[, vars_pcs]

dim(X)
dim(Y)

summary(X)

mod0 <- rda(Y ~ 1., X)  # Model with intercept only
mod1 <- rda(Y ~ ., X)  # Model with all explanatory variables

step.res <- ordiR2step(mod0, mod1, perm.max = 1000)

table = step.res$anova
table

table$ES.RDA = c(table$R2.adj[1], table$R2.adj[2]-table$R2.adj[1], 
                 table$R2.adj[3]-table$R2.adj[2], table$R2.adj[4]-table$R2.adj[3],
                 table$R2.adj[5]-table$R2.adj[4], table$R2.adj[6]-table$R2.adj[5],
                 table$R2.adj[7]-table$R2.adj[6], table$R2.adj[8]-table$R2.adj[7],
                 table$R2.adj[9]-table$R2.adj[8], table$R2.adj[10]-table$R2.adj[9],
                 table$R2.adj[11]-table$R2.adj[10], table$R2.adj[12]-table$R2.adj[11])
table = table[-12, ]

step.res$call

table

rownames(table) = c('1,25-(OH)2D', 'Race', 'Site', 'Antidepressent Use',
                    '25(OH)D', 'Total Starch Intake',  'Oral Antibiotic Use',  
                    'Physical Activity Score', 'Age',  'PPI')

# export
write.table(table, file='../data/RDA_beta_table.txt', sep='\t', col.names=NA)

library(ggplot2)
pdf('../figures/RDA_beta_uw.pdf')
covariates = rownames(table)
ggplot(table, aes(x=reorder(covariates, ES.RDA), y=ES.RDA, fill=covariates)) +
  labs(x = 'Non-redundant Covariants', y = 'Effect Size (Unweighted UniFrac)') +
  geom_bar(stat='identity') +
  scale_x_discrete(labels = c('PPI', 'Age', 'Physical Activity Score', 'Oral Antibiotic Use', 'Statin Use',
                              'Total Starch Intake', '25(OH)D', 'Antidepressant Use', 'Site', 'Race', expression('1,25(OH)'[2]*'D'))) +
  theme(axis.text=element_text(size=14, face='bold'), 
       axis.title=element_text(size=17,face="bold"), 
       legend.position="none") + 
  coord_flip()
dev.off()
