==> 3.2 Beta Diversity (UW) -- Permanova Adonis (Qiime1).ipynb <==
### This notebook is to perform permanova tests on unweighted UniFrac (qiime1 environment).

#### 1. Categorical Covariates (Permanova Test)

# Race (p-value = 0.001 ***)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c GIERACE \
-o ../beta-uw-results/unweighted-unifrac-GIERACE --method permanova

# Site (p-value = 0.001 ***)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c SITE \
-o ../beta-uw-results/unweighted-unifrac-SITE --method permanova 

# TUDRAMT (0.023)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c TUDRAMT \
-o ../beta-uw-results/unweighted-unifrac-TUDRAMT --method permanova 

# TURSMOKE (0.07)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c TURSMOKE \
-o ../beta-uw-results/unweighted-unifrac-TURSMOKE --method permanova 

# Antidepressant Use (p-value = 0.001 ***)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c M1ADEPR \
-o ../beta-uw-results/unweighted-unifrac-M1ADEPR --method permanova

# M1VITMND (0.709)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c M1VITMND \
-o ../beta-uw-results/unweighted-unifrac-M1VITMND --method permanova

# Antibiotic Use (p-value = 0.001 ***)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c M1ANTIB \
-o ../beta-uw-results/unweighted-unifrac-M1ANTIB --method permanova

# M1PROBI (p-value = 0.575)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c M1PROBI \
-o ../beta-uw-results/unweighted-unifrac-M1PROBI --method permanova 

# OHSEAS (p-value = 0.152)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c OHSEAS \
-o ../beta-uw-results/unweighted-unifrac-OHSEAS --method permanova 

# QLCOMP (p-value = 0.043)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c QLCOMP \
-o ../beta-uw-results/unweighted-unifrac-QLCOMP --method permanova 

# VDstatus (p-value = 0.796)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c VDstatus \
-o ../beta-uw-results/unweighted-unifrac-VDstatus --method permanova

# Statin Use (p-value = 0.01 **)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c M1STATIN \
-o ../beta-uw-results/unweighted-unifrac-M1STATIN --method permanova

# Antihistamine (0.407)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c Antihistamine \
-o ../beta-uw-results/unweighted-unifrac-Antihistamine --method permanova

# Laxative (p-value = 0.076)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c Laxative \
-o ../beta-uw-results/unweighted-unifrac-Laxative --method permanova

from statsmodels.sandbox.stats.multicomp import multipletests
multipletests([0.001,0.001, 0.023, 0.07, 0.001, 0.709, 0.001, 0.575, 
               0.152, 0.043, 0.796, 0.01, 0.407, 0.076],
               alpha=0.05, method='fdr_bh') 

#### 2. Continous Covariates (Adonis Test)

# Age (p-value = 0.03 *)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c Age \
-o ../beta-uw-results/unweighted-unifrac-Age --method adonis

# BMI (p-value = 0.009 **)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c BMI \
-o ../beta-uw-results/unweighted-unifrac-BMI --method adonis

# Pascore (p-value = 0.014 *)
compare_categories.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv \
-m ../data/mros_mapping_alpha.txt \
-c PASCORE \
-o ../beta-uw-results/unweighted-unifrac-PASCORE --method adonis

#### filter metadata based on vitD measurements

import numpy as np
import pandas as pd

mf = pd.read_csv('../data/mros_mapping_alpha.txt', sep='\t', index_col='#SampleID')

mf.head()

mf.shape

mf = mf.loc[:, ('OHVD3', 'OHV1D3', 'OHV24D3', 'ratio_activation', 'ratio_catabolism')]

mf.head()

mf.dropna(axis=0, how='any', inplace=True)

mf.to_csv('../data/mf.vitD.txt', sep='\t', index_label='#SampleID')

# filter the distance matrix based on filtered metadata file
filter_distance_matrix.py -i ../Qiime_updated/core-metrics-results/unweighted_unifrac_distance/distance-matrix.tsv  \
-o ../data/unweighted.dm.filtered.txt \
--sample_id_fp ../data/mf.vitd.ids.txt

# 1, 25 (p-value = 0.001 ***)
compare_categories.py -i ../data/unweighted.dm.filtered.txt \
-m ../data/mf.vitD.txt \
-c OHV1D3 \
-o ../beta-uw-results/unweighted-unifrac-OHV1D3 --method adonis

# 25 (p-value = 0.085)
compare_categories.py -i ../data/unweighted.dm.filtered.txt \
-m ../data/mf.vitD.txt \
-c OHVD3 \
-o ../beta-uw-results/unweighted-unifrac-OHVD3 --method adonis

# 24 (p-value = 0.002 **)
compare_categories.py -i ../data/unweighted.dm.filtered.txt \
-m ../data/mf.vitD.txt \
-c OHV24D3 \
-o ../beta-uw-results/unweighted-unifrac-OHV24D3 --method adonis

# ratio_act (p-value = 0.001 ***)
compare_categories.py -i ../data/unweighted.dm.filtered.txt \
-m ../data/mf.vitD.txt \
-c ratio_activation \
-o ../beta-uw-results/unweighted-unifrac-ratio_activation --method adonis

# ratio_cat (p-value = 0.001 ***)
compare_categories.py -i ../data/unweighted.dm.filtered.txt \
-m ../data/mf.vitD.txt \
-c ratio_catabolism \
-o ../beta-uw-results/unweighted-unifrac-ratio_catabolism --method adonis

from statsmodels.sandbox.stats.multicomp import multipletests
multipletests([0.03, 0.009, 0.014, 0.001, 0.085, 0.002, 0.001, 0.001], alpha=0.05, method='fdr_bh')
