==> 6.1 taxonomic correlation_1,25 (Python).ipynb <==
### This notebook is to use spearman correlation to check association between 1,25-(OH)2D and interested taxa

import warnings
warnings.filterwarnings("ignore")

import pandas as pd
import numpy as np
from scipy.stats import spearmanr, pearsonr
from statsmodels.sandbox.stats.multicomp import multipletests

import matplotlib.pylab as plt
%matplotlib inline

### merge taxonomy with mapping file

taxa = pd.read_csv('../data/RF_taxa_125.txt', sep='\t', index_col='#OTU ID')

print(taxa.shape)
taxa.head()

otu = pd.read_csv('../data/57316_mros_otus_rare_exp/57316_feature-table-rare.txt', sep='\t', skiprows=1, index_col='#OTU ID',dtype = str)
print(otu.shape)
otu.head()

# subset otu table with intersted ones only
otu_sub = otu.loc[taxa.index].transpose()
print(otu_sub.shape)
otu_sub.head()

mf = pd.read_csv('../data/mros_mapping_alpha.txt', sep='\t', index_col='#SampleID')

mf = mf[['OHV1D3', 'OHV24D3', 'OHVD3', 'ratio_activation', 'ratio_catabolism']]
print(mf.shape)
mf.head()

dat = pd.merge(mf, otu_sub, left_index=True, right_index=True)
print(dat.shape)
dat.head()

# convert to float type
dat = dat.apply(pd.to_numeric, errors='coerce')
dat.describe()

dat[dat.columns[7]].dtype

### correlation

otu_cols = dat.columns[mf.shape[1]:dat.shape[1]]
otu_cols

vars_vd = np.array(['OHV1D3', 'OHV24D3', 'OHVD3', 'ratio_activation', 'ratio_catabolism'])

results= []
i = 0
print(vars_vd[i])
for j in range(len(otu_cols)):
    tmp = dat[[vars_vd[i], otu_cols[j]]].dropna(axis=0, how='any')
    rho, pval = spearmanr(tmp[vars_vd[i]], tmp[otu_cols[j]])
    tax = taxa['Taxon'][otu_cols[j]]
    results.append([vars_vd[i], otu_cols[j], tax, rho, pval])
    results.append([vars_vd[i], otu_cols[j], rho, pval])

# output table    
results = pd.DataFrame(results, columns=['vars', 'otu', 'tax',
                                         'rho', 'pval']).dropna(axis=0, how='any')
results['fdr pval'] = multipletests(results['pval'], method = 'fdr_bh')[1]
results = results.sort_values(['fdr pval'], ascending=True)

# specific bacteria
index = results.loc[results['fdr pval'] <= 0.05].index
for i in range(len(index)):
    print(results.tax[index[i]], results['rho'][index[i]])

results

results.to_csv('../data/correlation_125.txt', sep='\t')

dat.head()

# check
dat.rename(columns={dat.columns[5]: taxa.Taxon[dat.columns[5]],
                    dat.columns[6]: taxa.Taxon[dat.columns[6]],
                    dat.columns[7]: taxa.Taxon[dat.columns[7]],
                    dat.columns[8]: taxa.Taxon[dat.columns[8]],
                    dat.columns[9]: taxa.Taxon[dat.columns[9]],
                    dat.columns[10]: taxa.Taxon[dat.columns[10]],
                    dat.columns[11]: taxa.Taxon[dat.columns[11]]}, inplace=True)
dat.head()

tmp = dat[['OHV1D3', 'k__Bacteria; p__Firmicutes; c__Clostridia; o__Clostridiales; f__; g__; s__']].dropna(axis=0, how='any')
spearmanr(tmp[tmp.columns[0]], tmp[tmp.columns[1]])

tmp = dat[['OHV1D3', 'k__Bacteria; p__Firmicutes; c__Clostridia; o__Clostridiales; f__Lachnospiraceae; g__; s__']].dropna(axis=0, how='any')
spearmanr(tmp[tmp.columns[0]], tmp[tmp.columns[1]])

tmp = dat[['OHV1D3', 'k__Bacteria; p__Firmicutes; c__Clostridia; o__Clostridiales; f__Christensenellaceae; g__; s__']].dropna(axis=0, how='any')
spearmanr(tmp[tmp.columns[0]], tmp[tmp.columns[1]])

tmp = dat[['OHV1D3', 'k__Bacteria; p__Firmicutes; c__Clostridia; o__Clostridiales; f__Ruminococcaceae; g__; s__']].dropna(axis=0, how='any')
spearmanr(tmp[tmp.columns[0]], tmp[tmp.columns[1]])

tmp = dat[['OHV1D3', 'k__Bacteria; p__Actinobacteria; c__Coriobacteriia; o__Coriobacteriales; f__Coriobacteriaceae; g__Collinsella; s__aerofaciens']].dropna(axis=0, how='any')
spearmanr(tmp[tmp.columns[0]], tmp[tmp.columns[1]])

tmp = dat[['OHV1D3', 'k__Bacteria; p__Bacteroidetes; c__Bacteroidia; o__Bacteroidales; f__Rikenellaceae; g__; s__']].dropna(axis=0, how='any')
spearmanr(tmp[tmp.columns[0]], tmp[tmp.columns[1]])
