==> 6.3 taxonomic correlation_Ratio Activation (Python).ipynb <==
### This notebook is to use spearman correlation to check association between Ratio_Activation and interested taxa

import warnings
warnings.filterwarnings("ignore")

import pandas as pd
import numpy as np
from scipy.stats import spearmanr, pearsonr
from statsmodels.sandbox.stats.multicomp import multipletests

import matplotlib.pylab as plt
%matplotlib inline

### merge taxonomy with mapping file

taxa = pd.read_csv('../data/RF_taxa_act.txt', sep='\t', index_col='#OTU ID')

print(taxa.shape)
taxa.head()

otu = pd.read_csv('../Qiime_updated/feature-table-rare5807.txt', sep='\t', skiprows=1, index_col='#OTU ID')
otu.head()

# combine taxa with otu
taxa = pd.merge(taxa, otu, left_index=True, right_index=True).transpose()
taxa.shape

taxa.head()

mf = pd.read_csv('../data/mros_mapping_alpha.txt', sep='\t', index_col='#SampleID')

mf = mf[['OHV1D3', 'OHV24D3', 'OHVD3', 'ratio_activation', 'ratio_catabolism', 'VDstatus']]
mf.head()

# need to substract the six otu abundance for each subject
sample_ids = mf.index
taxa = taxa.loc[sample_ids]
taxa.head()

dat = pd.merge(mf, taxa, left_index=True, right_index=True)
dat.head()

vars_vd = np.array(['OHVD3', 'OHV1D3', 'OHV24D3', 'ratio_activation', 'ratio_catabolism'])
dat[vars_vd] = dat[vars_vd].apply(pd.to_numeric, errors='coerce')

dat[vars_vd].describe()

### correlation

otu_cols = dat.columns[mf.shape[1]:dat.shape[1]]
len(otu_cols)

bt = pd.read_csv('../data/RF_taxa_act.txt', sep='\t', index_col='#OTU ID')
bt.head()

bt.index == dat.columns[mf.shape[1]:dat.shape[1]]

results= []
i = 3
print(vars_vd[i])
for j in range(len(otu_cols)):
    tmp = dat[[vars_vd[i], otu_cols[j]]].dropna(axis=0, how='any')
    rho, pval = spearmanr(tmp[vars_vd[i]], tmp[otu_cols[j]])
    tax = bt['Taxon'][otu_cols[j]]
    results.append([vars_vd[i], otu_cols[j], tax, rho, pval])
    results.append([vars_vd[i], otu_cols[j], rho, pval])

# output table    
results = pd.DataFrame(results, columns=['vars', 'otu', 'tax',
                                         'rho', 'pval']).dropna(axis=0, how='any')
results['fdr pval'] = multipletests(results['pval'], method = 'fdr_bh')[1]
results = results.sort_values(['fdr pval'], ascending=True)

# specific bacteria
index = results.loc[results['fdr pval'] <= 0.05].index
for i in range(len(index)):
    print(results.tax[index[i]], results['fdr pval'][index[i]])

# check
results

results.to_csv('../data/correlation_Act.txt', sep='\t')

## double check the results

dat.rename(columns={dat.columns[6]: bt.Taxon[dat.columns[6]],
                    dat.columns[7]: bt.Taxon[dat.columns[7]],
                    dat.columns[8]: bt.Taxon[dat.columns[8]],
                    dat.columns[9]: bt.Taxon[dat.columns[9]],
                    dat.columns[10]: bt.Taxon[dat.columns[10]],
                    dat.columns[11]: bt.Taxon[dat.columns[11]],
                    dat.columns[12]: bt.Taxon[dat.columns[12]],
                    dat.columns[13]: bt.Taxon[dat.columns[13]]}, inplace=True)
dat.head()

tmp = dat[['ratio_activation', 'k__Bacteria; p__Firmicutes; c__Clostridia; o__Clostridiales; f__Ruminococcaceae; g__; s__']].dropna(axis=0, how='any')
spearmanr(tmp[tmp.columns[0]], tmp[tmp.columns[1]])

tmp = dat[['ratio_activation', 'k__Bacteria; p__Firmicutes; c__Clostridia; o__Clostridiales; f__Ruminococcaceae; g__Oscillospira; s__']].dropna(axis=0, how='any')
spearmanr(tmp[tmp.columns[0]], tmp[tmp.columns[1]])

tmp = dat[['ratio_activation', 'k__Bacteria; p__Firmicutes; c__Clostridia; o__Clostridiales; f__Lachnospiraceae; g__Dorea; s__']].dropna(axis=0, how='any')
spearmanr(tmp[tmp.columns[0]], tmp[tmp.columns[1]])

tmp = dat[['ratio_activation', 'k__Bacteria; p__Firmicutes; c__Clostridia; o__Clostridiales; f__Clostridiaceae; g__; s__']].dropna(axis=0, how='any')
spearmanr(tmp[tmp.columns[0]], tmp[tmp.columns[1]])
